pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["Anastasia <aastavitskaya@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.3"
djangorestframework-camel-case = "^1.4.2"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.31.0"
mixer = "^7.2.2"
pytest = "^8.0.0"
pytest-mock = "^3.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
